agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnspython>=2.8.0",
    "ping3>=5.1.5",
    "redis>=7.0.0",
    "requests>=2.32.5",
]
